TWINFACES-576 refactor: move skeletons to features
