+++ src/app/core/twinclass/loading.tsx
@@ -1,4 +1,4 @@
-import { TableSkeleton } from "@/widgets/skeletons";
+import { TableSkeleton } from "@/features/ui/skeletons";
 
 export default function Loading() {
   return <TableSkeleton />;
