+++ src/features/ui/skeletons/carousel.tsx
