+++ src/features/ui/skeletons/index.ts
